pyproject.toml
[tool.poetry]
name = "resma-cli"
version = "0.1.0"
description = ""
authors = ["Thiago Campos <commit@thigcampos.com>", "Ivan Santiago <ivansantiago.junior@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
resma = "resma_cli.main:app"

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.12.3"
jinja2 = "^3.1.4"
markdown = "^3.6"
python-frontmatter = "^1.1.0"
toml = "^0.10.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.12.2"
ruff = "^0.5.6"
mypy = "^1.11.1"


[tool.pytest.ini_options]
pythonpath = "."

[tool.isort]
profile = "black"
line_length = 79

[tool.mypy]
ignore_missing_imports = true
strict = false
exclude = "tests"

[tool.ruff]
target-version = "py312"
line-length = 79

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']
ignore = ['PLR0913', 'PLR0917']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
pre_test = 'task lint'
test = 'pytest -s -x --cov=resma_cli -vv'
post_test = 'coverage html'
mypy = "mypy . --ignore-missing-imports --check-untyped-defs"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
